perf(www): defer pricing accordion runtime
