+++ apps/www/app/[locale]/(app)/(shared)/(site)/page.tsx
@@ -6,7 +6,7 @@ import type { Locale } from "next-intl";
 import { getTranslations } from "next-intl/server";
 import { Community } from "@/components/marketing/about/community";
 import { Curricula } from "@/components/marketing/about/curricula";
-import { Faq } from "@/components/marketing/about/faq";
+import { Faq } from "@/components/marketing/about/faq/landing";
 import { Features } from "@/components/marketing/about/features";
 import { Hero } from "@/components/marketing/about/hero";
 import { Logos } from "@/components/marketing/about/logos";
